split off validation into its own type
